+++ src/Gateway/Response/Mapper/HostedPaymentPageMapper.js
@@ -9,7 +9,11 @@ class HostedPaymentPageMapper extends AbstractMapper {
         const values = {
             mid: typeof this.source.mid !== 'undefined' ? this.source.mid : null,
             forwardUrl: typeof this.source.forwardUrl !== 'undefined' ? this.source.forwardUrl : null,
-            order: typeof this.source.order !== 'undefined' ? new OrderMapper(this.source.order).mappedObject : null
+            order: typeof this.source.order !== 'undefined' ? new OrderMapper(this.source.order).mappedObject : null,
+            numberRetry:
+                typeof this.source.numberRetry !== 'undefined'
+                ? this.source.numberRetry
+                : null
         };
 
         this._modelObject = new Transaction(values);