+++ src/test/unit/Gateway/HTTP/SimpleHTTPClient.test.js
@@ -73,7 +73,7 @@ describe('Test SimpleHTTPClient constructor, getters & setters', () => {
 describe('Test order request method', () => {
     const SimpleHTTPClient = require('../../../../Gateway/HTTP/SimpleHTTPClient');
 
-    it('calls API if everything is empty', async (done) => {
+    it('calls API if everything is empty', async () => {
         let mockConfig = new Configuration({});
         mockConfig.timeout = 5;
         mockConfig.apiEndpoint = '{API_ENDPOINT}';
@@ -99,6 +99,7 @@ describe('Test order request method', () => {
         expect(axios.request).toHaveBeenCalledWith({
             url: '{ENDPOINT}',
             method: 'POST',
+            baseURL: '',
             timeout: 5000,
             headers: {
                 Accept: '{HTTP_ACCEPT}',
@@ -106,14 +107,12 @@ describe('Test order request method', () => {
                 Authorization: 'Bearer {API_TOKEN}',
                 'Content-Type': 'application/x-www-form-urlencoded'
             },
-            baseURL: '',
             data: ''
         });
 
-        done();
     });
 
-    it('calls API if everything is right', async (done) => {
+    it('calls API if everything is right', async () => {
         let mockConfig = new Configuration({});
         mockConfig.timeout = 5;
         mockConfig.apiEndpoint = '{API_ENDPOINT}';
@@ -155,10 +154,9 @@ describe('Test order request method', () => {
             data: 'mockProp=mockValue&mockProp2=mockValue2'
         });
 
-        done();
     });
 
-    it('calls API with additional headers if everything is right', async (done) => {
+    it('calls API with additional headers if everything is right', async () => {
         let mockConfig = new Configuration({});
         mockConfig.timeout = 5;
         mockConfig.apiEndpoint = '{API_ENDPOINT}';
@@ -205,10 +203,9 @@ describe('Test order request method', () => {
             data: 'mockProp=mockValue&mockProp2=mockValue2'
         });
 
-        done();
     });
 
-    it('calls API in GET if everything is right', async (done) => {
+    it('calls API in GET if everything is right', async () => {
         let mockConfig = new Configuration({});
         mockConfig.timeout = 5;
         mockConfig.apiEndpoint = '{API_ENDPOINT}';
@@ -248,10 +245,9 @@ describe('Test order request method', () => {
             }
         });
 
-        done();
     });
 
-    it('calls API with proxy if everything is right', async (done) => {
+    it('calls API with proxy if everything is right', async () => {
         let mockConfig = new Configuration({});
         mockConfig.timeout = 5;
         mockConfig.apiEndpoint = '{API_ENDPOINT}';
@@ -300,10 +296,9 @@ describe('Test order request method', () => {
             data: 'mockProp=mockValue&mockProp2=mockValue2'
         });
 
-        done();
     });
 
-    it('calls API with basic credentials if everything is right', async (done) => {
+    it('calls API with basic credentials if everything is right', async () => {
         let mockConfig = new Configuration({});
         mockConfig.timeout = 5;
         mockConfig.apiEndpoint = '{API_ENDPOINT}';
@@ -346,10 +341,9 @@ describe('Test order request method', () => {
             data: 'mockProp=mockValue&mockProp2=mockValue2'
         });
 
-        done();
     });
 
-    it('calls data API if everything is right', async (done) => {
+    it('calls data API if everything is right', async () => {
         let mockConfig = new Configuration({});
         mockConfig.timeout = 5;
         mockConfig.apiEndpoint = '{API_ENDPOINT}';
@@ -383,20 +377,21 @@ describe('Test order request method', () => {
             url: '{ENDPOINT}',
             method: 'POST',
             baseURL: '{DATA_API_ENDPOINT}',
-            timeout: 60000,
+            timeout: 10000,
             headers: {
                 Accept: '{HTTP_ACCEPT}',
                 'User-Agent': '{DATA_HTTP_USER_AGENT}',
                 'X-Who-Api': '{DATA_HTTP_USER_AGENT}',
                 'Content-Type': 'application/json'
             },
+            httpAgent: expect.objectContaining({ keepAlive: false }),
+            httpsAgent: expect.objectContaining({ keepAlive: false }),
             data: '{"mockProp":"mockValue","mockProp2":"mockValue2"}'
         });
 
-        done();
     });
 
-    it('calls data API with error without stopping request', async (done) => {
+    it('calls data API with error without stopping request', async () => {
         let mockConfig = new Configuration({});
         mockConfig.timeout = 5;
         mockConfig.apiEndpoint = '{API_ENDPOINT}';
@@ -421,10 +416,9 @@ describe('Test order request method', () => {
 
         await expect(client.request('POST', '{ENDPOINT}', { body: params, baseUrl: '{API_ENDPOINT}', isData: true })).resolves.not.toThrow();
 
-        done();
     });
 
-    it('errors if method is not string', async (done) => {
+    it('errors if method is not string', async () => {
         let mockConfig = new Configuration({});
         let client = new SimpleHTTPClient(mockConfig);
 
@@ -441,10 +435,9 @@ describe('Test order request method', () => {
 
         expect(InvalidArgumentException).toHaveBeenCalledWith('HTTP METHOD must a string and a valid HTTP METHOD Value');
 
-        done();
     });
 
-    it('errors if method is not valid', async (done) => {
+    it('errors if method is not valid', async () => {
         let mockConfig = new Configuration({});
         let client = new SimpleHTTPClient(mockConfig);
 
@@ -461,10 +454,9 @@ describe('Test order request method', () => {
 
         expect(InvalidArgumentException).toHaveBeenCalledWith('HTTP METHOD must a string and a valid HTTP METHOD Value');
 
-        done();
     });
 
-    it('errors if endpoint is not string', async (done) => {
+    it('errors if endpoint is not string', async () => {
         let mockConfig = new Configuration({});
         let client = new SimpleHTTPClient(mockConfig);
 
@@ -481,10 +473,9 @@ describe('Test order request method', () => {
 
         expect(InvalidArgumentException).toHaveBeenCalledWith('Endpoint must be a string and a valid api endpoint');
 
-        done();
     });
 
-    it('errors if endpoint is empty', async (done) => {
+    it('errors if endpoint is empty', async () => {
         let mockConfig = new Configuration({});
         let client = new SimpleHTTPClient(mockConfig);
 
@@ -501,10 +492,9 @@ describe('Test order request method', () => {
 
         expect(InvalidArgumentException).toHaveBeenCalledWith('Endpoint must be a string and a valid api endpoint');
 
-        done();
     });
 
-    it('errors if order API errors', async (done) => {
+    it('errors if order API errors', async () => {
         let mockConfig = new Configuration({});
         mockConfig.timeout = 5;
         mockConfig.apiEndpoint = '{API_ENDPOINT}';
@@ -545,10 +535,9 @@ describe('Test order request method', () => {
 
         expect(ApiErrorException).toHaveBeenCalledWith('{ERROR_MESSAGE}', '{ERROR_STATUS}', '{ERROR_CODE}', '{ERROR_DESCRIPTION}');
 
-        done();
     });
 
-    it('errors if order API errors with no info', async (done) => {
+    it('errors if order API errors with no info', async () => {
         let mockConfig = new Configuration({});
         mockConfig.timeout = 5;
         mockConfig.apiEndpoint = '{API_ENDPOINT}';
@@ -604,10 +593,9 @@ describe('Test order request method', () => {
             });
         }
 
-        done();
     });
 
-    it('errors if axios errors', async (done) => {
+    it('errors if axios errors', async () => {
         let mockConfig = new Configuration({});
         mockConfig.timeout = 5;
         mockConfig.apiEndpoint = '{API_ENDPOINT}';
@@ -636,6 +624,5 @@ describe('Test order request method', () => {
             expect(err.message).toEqual('Sample error');
         }
 
-        done();
     });
 });